Generate common application service for tests
